+++ lib/slack/web/api/endpoints/auth.rb
@@ -6,6 +6,17 @@ module Slack
       module Endpoints
         module Auth
           #
+          # This method revokes an access token. Use it when you no longer need a token. For example, with a Sign In With Slack app, call this to log a user out.
+          #
+          # @option options [Object] :test
+          #   Setting this parameter to 1 triggers a testing mode where the specified token will not actually be revoked.
+          # @see https://api.slack.com/methods/auth.revoke
+          # @see https://github.com/dblock/slack-api-ref/blob/master/methods/auth/auth.revoke.json
+          def auth_revoke(options = {})
+            post('auth.revoke', options)
+          end
+
+          #
           # This method checks authentication and tells you who you are.
           #
           # @see https://api.slack.com/methods/auth.test